Struct merlin::TranscriptRng

pub struct TranscriptRng { /* fields omitted */ }

An RNG providing synthetic randomness to the prover.

A TranscriptRng is constructed from a Transcript using a TranscriptRngBuilder; see its documentation for details on how to construct one.

The transcript RNG construction is described in the Generating Randomness section of the Merlin website.