Crate merlin

Note that docs will only build on nightly Rust until RFC 1990 stabilizes.

Structs

Transcript

A transcript of a public-coin argument.

TranscriptRng

An RNG providing synthetic randomness to the prover.

TranscriptRngBuilder

Constructs a TranscriptRng by rekeying the Transcript with prover secrets and an external RNG.